+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/rating/framework/MetaRating.java
@@ -39,4 +39,8 @@ public interface MetaRating extends Rating, RatingTranslator, MetaRatingMerger {
     default MetaRating asMetaRating() {
         return this;
     }
+
+    RatingTranslator translator();
+
+    MetaRatingMerger merger();
 }